requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
brotlipy=0.7.0=py38h9ed2024_1003
ca-certificates=2020.12.8=hecd8cb5_0
certifi=2020.12.5=py38hecd8cb5_0
cffi=1.14.4=py38h2125817_0
chardet=4.0.0=py38hecd8cb5_1003
click=7.1.2=py_0
cryptography=3.3.1=py38hbcfaee0_0
dataclasses=0.7=py38_0
filelock=3.0.12=py_0
freetype=2.10.4=ha233b18_0
idna=2.10=py_0
intel-openmp=2019.4=233
joblib=1.0.0=pyhd3eb1b0_0
jpeg=9b=he5867d9_2
lcms2=2.11=h92f6f08_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libpng=1.6.37=ha441bb4_0
libprotobuf=3.13.0.1=hab81aa3_0
libtiff=4.1.0=hcb84e12_1
libuv=1.40.0=haf1e3a3_0
lz4-c=1.9.2=h79c402e_3
mkl=2019.4=233
mkl-service=2.3.0=py38h9ed2024_0
mkl_fft=1.2.0=py38hc64f4ea_0
mkl_random=1.1.1=py38h959d312_0
ncurses=6.2=h0a44026_1
ninja=1.10.2=py38hf7b0b51_0
numpy=1.19.2=py38h456fd55_0
numpy-base=1.19.2=py38hcfb5961_0
olefile=0.46=py_0
openssl=1.1.1i=h9ed2024_0
packaging=20.8=pyhd3eb1b0_0
pillow=8.0.1=py38h5270095_0
pip=20.3.3=py38hecd8cb5_0
protobuf=3.13.0.1=py38hb1e8313_1
pycparser=2.20=py_2
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=py_0
pysocks=1.7.1=py38_1
python=3.8.5=h26836e1_1
python_abi=3.8=1_cp38
pytorch=1.7.1=py3.8_0
readline=8.0=h1de35cc_0
regex=2020.11.13=py38h9ed2024_0
requests=2.25.1=pyhd3eb1b0_0
sacremoses=master=py_0
setuptools=51.0.0=py38hecd8cb5_2
six=1.15.0=py38hecd8cb5_0
sqlite=3.33.0=hffcf06c_0
tk=8.6.10=hb0a8c7a_0
tokenizers=0.9.4=py38_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cpu
tqdm=4.54.1=pyhd3eb1b0_0
transformers=4.1.1=pyh39e3cac_0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.2=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
